cli/pyproject.toml
[tool.poetry]
name = "robo-cli"
version = "0.1.0"
description = ""
authors = ["Ossi Rajuvaara <ossi@robocorp.com>"]
readme = "README.md"
packages = [{include = "robo_cli"}]

[tool.poetry.scripts]
robo = "robo_cli.app:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.7.0"
rich = "^13.3.2"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ruff = "^0.0.255"
mypy = "^1.1.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
